pyproject.toml
[project]
name = "realtimetranslation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-speech>=2.33.0",
    "numpy>=2.3.1",
    "scipy>=1.16.0",
    "sounddevice>=0.5.2",
]
